Fix exception when closing empty path
It happens when you open Android studio's layout preview where some of your views try to use IconicsDrawable. Because layout preview doesn't support Paint.getTextPath(), path itself will be empty, and when you try to close it, it throws the exception.

Simple fix is to check if path is not empty before trying to close it.